:sparkles: Séparation en composants
